Fix: Corrección de la implementación. Ahora la lógica esta en el Service. Ahora la documentación está actualizada

+++ src/main/java/com/ccsw/tutorial/author/AuthorServiceImpl.java
@@ -6,6 +6,7 @@ import com.ccsw.tutorial.author.model.AuthorSearchDto;
 import com.ccsw.tutorial.common.deleteCheck.DeleteCheckObject;
 import com.ccsw.tutorial.common.deleteCheck.DeleteCheckResponseDto;
 import com.ccsw.tutorial.exceptions.NoIdFoundException;
+import com.ccsw.tutorial.exceptions.NotDeleteableException;
 import com.ccsw.tutorial.game.GameRepository;
 import com.ccsw.tutorial.game.model.Game;
 import jakarta.transaction.Transactional;
@@ -73,12 +74,15 @@ public class AuthorServiceImpl implements AuthorService {
      * {@inheritDoc}
      */
     @Override
-    public void delete(Long id) throws NoIdFoundException {
+    public void delete(Long id) throws NoIdFoundException, NotDeleteableException {
 
         if (this.get(id) == null) {
             throw new NoIdFoundException();
         }
-
+        DeleteCheckResponseDto deleteCheck = isDeleteable(id);
+        if(!deleteCheck.isCanDelete()){
+            throw new NotDeleteableException(deleteCheck.getReason());
+        }
         this.authorRepository.deleteById(id);
     }
